Move Partner enum to dockstore-common
