+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/CloudInstance.java
@@ -1,6 +1,7 @@
 package io.dockstore.webservice.core;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import io.dockstore.common.Partner;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
